# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Asset Tracker API",
    "version": "2.0.0",
    "description": "Comprehensive asset management system API with multi-tenancy, RBAC, webhooks, and advanced features.",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Session token from NextAuth authentication"
      }
    },
    "schemas": {
...
      "AssetReservation": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "assetId": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["pending", "approved", "rejected", "cancelled", "completed"] },
          "notes": { "type": "string", "nullable": true },
          "approvedBy": { "type": "string", "format": "uuid", "nullable": true },
          "approvedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "assetId", "userId", "startDate", "endDate", "status"]
      },
...
      "Asset": {
        "type": "object",
        "properties": {
          "assetid": { "type": "string", "format": "uuid" },
          "assetname": { "type": "string" },
          "assettag": { "type": "string" },
          "serialnumber": { "type": "string" },
          "modelid": { "type": "string", "format": "uuid", "nullable": true },
          "specs": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true },
          "purchaseprice": { "type": "number", "nullable": true },
          "purchasedate": { "type": "string", "format": "date", "nullable": true },
          "mobile": { "type": "boolean", "nullable": true },
          "requestable": { "type": "boolean", "nullable": true },
          "organizationId": { "type": "string", "format": "uuid", "nullable": true }
        },
        "required": ["assetid", "assetname", "assettag", "serialnumber"]
      },
...
    }
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/reservations": {
      "get": {
        "tags": ["Reservations"],
        "summary": "List asset reservations",
        "parameters": [
          { "name": "assetId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "userId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "List of reservations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AssetReservation" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Reservations"],
        "summary": "Create a reservation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "assetId": { "type": "string", "format": "uuid" },
                  "startDate": { "type": "string", "format": "date-time" },
                  "endDate": { "type": "string", "format": "date-time" },
                  "notes": { "type": "string" }
                },
                "required": ["assetId", "startDate", "endDate"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Reservation created" },
          "400": { "description": "Asset already reserved or validation error" }
        }
      }
    },
...
    "/asset": {
      "get": {
        "tags": ["Assets"],
        "summary": "List all assets",
        "parameters": [
          { "name": "id", "in": "query", "schema": { "type": "string", "format": "uuid" }, "description": "Get specific asset by ID" }
        ],
        "responses": {
          "200": {
            "description": "List of assets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Asset" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Assets"],
        "summary": "Create a new asset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Asset" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Asset created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Asset" }
              }
            }
          },
          "400": { "description": "Validation error" }
        }
      },
      "put": {
        "tags": ["Assets"],
        "summary": "Update an asset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Asset" }
            }
          }
        },
        "responses": {
          "200": { "description": "Asset updated" },
          "400": { "description": "assetid required" }
        }
      }
    }
  }
}